reactor by removing unnecessary redundancy
